requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-default-group==1.2.4
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.0
easydict==1.13
flatbuffers==25.9.23
fonttools==4.60.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
jax==0.7.1
jaxlib==0.7.1
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.7
mediapipe==0.10.18
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
onnx==1.19.1
onnxruntime-gpu==1.23.2
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==10.3.0
pluggy==1.6.0
prettytable==3.16.0
protobuf==4.25.8
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.13.0
sentencepiece==0.2.1
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
sounddevice==0.5.3
sqlite-fts4==1.0.3
sqlite-utils==3.36
stringzilla==4.2.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tqdm==4.66.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
wheel==0.45.1
